src/requirements.txt
black~=22.0
flake8>=3.7.9, <4.0
ipython>=7.31.1, <8.0
isort~=5.0
jupyter~=1.0
jupyterlab~=3.0
pillow~=8.0
kedro[pandas.CSVDataSet]==0.18.0
kedro-telemetry~=0.2.0
nbstripout~=0.4
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=6.2
s3fs==2022.1.0 
boto3==1.20.24 
aiobotocore==2.1.2 